Added node_modules to Sankhadeep's module, changed Ananya_Arpita Module's flask app file to app.py, updated the readme.md, added the data folder to Piush_Ravi module
